Move Loading to shared components
