+++ x-pack/plugins/enterprise_search/public/applications/shared/loading/loading.scss
